🧰❄️: move progress indicator into separate module
